feat:db contoller and serve dynamic data
