+++ cubic-workshop/config/database.json
@@ -1,1 +1,1 @@
-[{"id":"3d452c36-9dfb-48ff-b04a-f5a2d7dbef5d","name":"Default","description":"This is default cube","imageUrl":"https://ae01.alicdn.com/kf/HTB1CSddXRxRMKJjy0Fdq6yifFXa6/Gan-356-Air-SM-3x3-Black-Magic-cube-GAN-Air-SM-Magnetic-3x3x3-Speed-cube-gans.jpg","difficulty":1},{"id":"d85b65ef-b39a-49d8-a6a3-277cc05bc780","name":"Default1","description":"This is default1 cube","imageUrl":"https://thingsidesire.com/wp-content/uploads/2018/06/Eco-Dark-Rubik%E2%80%99s-Cube2.jpg","difficulty":2},{"id":"4a4a4aab-a8bd-4417-b654-97627ea10a71","name":"Default2","description":"This is default2 cube","imageUrl":"https://images-na.ssl-images-amazon.com/images/I/61izOzq%2BBAL._SY355_.jpg","difficulty":3},{"id":"ffebafe3-fe14-4ce7-8ef1-cc2a33541eac","name":"Default3","description":"This is default3 cube","imageUrl":"https://images-na.ssl-images-amazon.com/images/I/61HpQqVQ37L._SY355_.jpg","difficulty":4}]
+[{"id":"3d452c36-9dfb-48ff-b04a-f5a2d7dbef5d","name":"Default","description":"This is default cube","imageUrl":"https://ae01.alicdn.com/kf/HTB1CSddXRxRMKJjy0Fdq6yifFXa6/Gan-356-Air-SM-3x3-Black-Magic-cube-GAN-Air-SM-Magnetic-3x3x3-Speed-cube-gans.jpg","difficulty":1},{"id":"d85b65ef-b39a-49d8-a6a3-277cc05bc780","name":"Default1","description":"This is default1 cube","imageUrl":"https://thingsidesire.com/wp-content/uploads/2018/06/Eco-Dark-Rubik%E2%80%99s-Cube2.jpg","difficulty":2},{"id":"4a4a4aab-a8bd-4417-b654-97627ea10a71","name":"Default2","description":"This is default2 cube","imageUrl":"https://images-na.ssl-images-amazon.com/images/I/61izOzq%2BBAL._SY355_.jpg","difficulty":3},{"id":"ffebafe3-fe14-4ce7-8ef1-cc2a33541eac","name":"Default3","description":"This is default3 cube","imageUrl":"https://images-na.ssl-images-amazon.com/images/I/61HpQqVQ37L._SY355_.jpg","difficulty":4},null]
